Moving showDay helper methods to UserPreferences.js
